Extract Metrics to libs
