+++ extension-manifest-v2/package.json
@@ -48,8 +48,9 @@
   },
   "homepage": "https://github.com/ghostery/ghostery-extension#readme",
   "dependencies": {
-    "@ghostery/ui": "../ui",
     "@cliqz/url-parser": "^1.1.4",
+    "@ghostery/libs": "../libs",
+    "@ghostery/ui": "../ui",
     "@sentry/browser": "7.1.1",
     "@whotracksme/webextension-packages": "^0.3.1",
     "classnames": "^2.2.5",
