+++ extension-manifest-v2/src/classes/Account.js
@@ -21,7 +21,7 @@ import conf from './Conf';
 import dispatcher from './Dispatcher';
 import { alwaysLog, log } from '../utils/common';
 import Api from '../utils/api';
-import metrics from './Metrics';
+import metrics from './MetricsWrapper';
 import ghosteryDebugger from './Debugger';
 import { cookiesGet, setAllLoginCookies, cookiesRemove } from '../utils/cookies';
