+++ libs/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@ghostery/libs",
+  "version": "1.0.0",
+  "private": true,
+  "description": "",
+  "type": "module",
+  "license": "MPL-2.0",
+  "exports": {
+    ".": "./src/index.js"
+  }
+}
