+++ libs/src/index.js
@@ -1,0 +1,3 @@
+import Metrics from './metrics';
+
+export { Metrics };
